PIDWrapper: Changed floats to doubles

+++ src/PIDWrapper.cpp
@@ -68,10 +68,10 @@ foilboat_controller::FoilboatControl PIDWrapper::control(
 {
   foilboat_controller::FoilboatControl control_out;
   ROS_INFO("========== Control loop ==========");
-  float flap_control = 0;
+  double flap_control = 0;
   if(target->altitudeTarget > 0)
   {
-    float altitude_rate_target = altitude_controller.update(target->altitudeTarget, state->altitude, time);
+    double altitude_rate_target = altitude_controller.update(target->altitudeTarget, state->altitude, time);
     ROS_INFO("Altitude rate target: %f, Altitude target: %f, Altitude state: %f", altitude_rate_target, target->altitudeTarget, state->altitude);
     control_out.altitudeRateTarget = altitude_rate_target;
     flap_control = -altitude_rate_controller.update(altitude_rate_target, state->altitudeRate, time);
@@ -79,13 +79,13 @@ foilboat_controller::FoilboatControl PIDWrapper::control(
   }
 
   // Manual elevator control:
-  float elevator_control = target->elevatorTarget;
+  double elevator_control = target->elevatorTarget;
 
   // Pitch angle control:
   // float pitch_control = -pitch_controller.update(target_pitch, state->pitch, time) * 30 * 3.14 / 180;
   // ROS_INFO("Pitch control: %f, Pitch target: %f, Pitch state: %f, time: %f", pitch_control, target_pitch, state->pitch, time);
 
-  float roll_control = roll_controller.update(target->rollTarget, state->roll, time);
+  double roll_control = roll_controller.update(target->rollTarget, state->roll, time);
   ROS_INFO("Roll control: %f, Roll target: %f, Roll state: %f", roll_control, target->rollTarget, state->roll);
   control_out.rightFoil = roll_control + flap_control - state->pitch;
   control_out.leftFoil = -roll_control + flap_control - state->pitch;
